Wallet UI improvement.

+++ CashCare/Models/Wallet/Saving.cs
@@ -5,7 +5,7 @@ namespace CashCare.Models.Wallet
     public class Saving
     {
         public int Id { get; set; }
-        public int MonthlySavingAmount { get; set; }
+        public decimal MonthlySavingAmount { get; set; }
 
         [ForeignKey("Wallet")]
         public int WalletId { get; set; }
